Separated pandas backend into a singleton (#333)
